+++ lib/osm.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/area.dart';
 export 'src/bounds.dart';
+export 'src/edit.dart';
 export 'src/element.dart';
 export 'src/exception.dart';
 export 'src/filter.dart';